drf_api/requirements.txt
django>=4.0,<5.0
djangorestframework>=3.12,<4.0
pyyaml>=5.0
requests>=2.20
django-cors-headers>=3.10
